moved files to correct locations updated login to be a mat-card
