+++ src/app/Components/login/login.component.css
@@ -1,5 +1,4 @@
 .form-full-width {
-    min-width: 150px;
     max-width: 500px;
-    width:100%;
+    width: 100%;
 }
