+++ src/app/app.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { AuthService } from '../app/auth.service';
+import { AuthService } from './Services/auth.service';
 
 @Component({
   selector: 'app-root',
