--- src/app/login/login.component.html
@@ -1,18 +1,0 @@
-<mat-toolbar>
-  <span>Login</span>
-</mat-toolbar>
-
-<mat-card>
-  Domain of Amy-Core
-  <div class="form-group">
-    <mat-form-field class="form-full-width">
-      <input matInput [(ngModel)]="inputText" type="text" class="form-control" id="url" placeholder="http://localhost:8080/rest/"
-        required>
-    </mat-form-field>
-  </div>
-
-  <button mat-raised-button type="submit" color="primary" (click)="login()">Login </button>
-
-
-
-</mat-card>
